Find actual first item in input for TintedShapedRecipe for copying tint component
The input always gets populated with width*height, including empty stacks. Wasn't detectable with slabs as they always take a 3*1 row and buttons/logsToPlanks as they take 1*1. With a mirrored stairs recipe though, it takes a full 3x3 box and the first is empty, so 'copyComponentsToNewStack' returned an empty stack.